add check types and its omponents
