+++ uikit/domql/Check/indicator.js
@@ -1,28 +1,25 @@
 'use strict'
 
-import { Icon } from '@symbo.ls/icon'
-
 export const CheckIndicator = {
-  checkIcon: {
-    extend: Icon,
-    props: { Icon: { name: 'check' } }
-  },
   props: {
     padding: 'V',
     boxSize: 'fit-content fit-content',
     round: '100%',
-    background: '#0474F2',
-    checkIcon: { fontSize: 'X' }
+    background: '#0474F2'
+  },
+  Icon: {
+    fontSize: 'X',
+    name: 'check'
   }
 }
 
-export const CheckIconWIthBorder = {
+export const CheckIndicatorWithBorder = {
   extend: CheckIndicator,
   props: {
     border: '2px solid #04F214',
     color: '#04F214',
     padding: 'Y+V',
     background: 'transparent',
-    checkIcon: { fontSize: 'G' }
+    Icon: { fontSize: 'G' }
   }
 }
