+++ uikit/domql/Check/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@symbo.ls/radiobutton",
+  "name": "@symbo.ls/check",
   "version": "2.11.132",
   "main": "index.js",
   "license": "MIT",
