+++ uikit/domql/Form/index.js
@@ -1,8 +1,5 @@
 'use strict'
 
-export * from './Checkbox'
-export * from './Radio'
-export * from './ToggleSwitch'
 export * from './FieldLabel'
 export * from './ContactForm'
 export * from './SignUpForm'
