+++ uikit/domql/index.js
@@ -5,10 +5,7 @@
 export * from '@symbo.ls/atoms'
 export * from '@symbo.ls/box'
 export * from '@symbo.ls/indicatordot'
-export * from '@symbo.ls/checkindicator'
-export * from '@symbo.ls/checkbox'
-export * from '@symbo.ls/radiobutton'
-export * from '@symbo.ls/toggleswitch'
+export * from '@symbo.ls/check'
 export * from '@symbo.ls/titleparagraph'
 export * from '@symbo.ls/notificationindicator'
 export * from '@symbo.ls/field'
